fix: HRA Exemption calculation in case of multiple salary structure assignments

+++ erpnext/payroll/doctype/salary_structure/test_salary_structure.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors and Contributors
 # See license.txt
 
-
 import frappe
 from frappe.tests.utils import FrappeTestCase
 from frappe.utils import add_years, date_diff, get_first_day, nowdate
@@ -164,12 +163,10 @@ def make_salary_structure(
 	other_details=None,
 	test_tax=False,
 	company=None,
-	currency=None,
+	currency=erpnext.get_default_currency(),  # noqa
 	payroll_period=None,
 	include_flexi_benefits=False,
 ):
-	currency = erpnext.get_default_currency()
-
 	if test_tax:
 		frappe.db.sql("""delete from `tabSalary Structure` where name=%s""", (salary_structure))
 
@@ -229,12 +226,14 @@ def create_salary_structure_assignment(
 	salary_structure,
 	from_date=None,
 	company=None,
-	currency=None,
+	currency=erpnext.get_default_currency(),  # noqa
 	payroll_period=None,
+	base=None,
+	allow_duplicate=False,
 ):
-	currency = erpnext.get_default_currency()
-
-	if frappe.db.exists("Salary Structure Assignment", {"employee": employee}):
+	if not allow_duplicate and frappe.db.exists(
+		"Salary Structure Assignment", {"employee": employee}
+	):
 		frappe.db.sql("""delete from `tabSalary Structure Assignment` where employee=%s""", (employee))
 
 	if not payroll_period:
@@ -247,7 +246,7 @@ def create_salary_structure_assignment(
 
 	salary_structure_assignment = frappe.new_doc("Salary Structure Assignment")
 	salary_structure_assignment.employee = employee
-	salary_structure_assignment.base = 50000
+	salary_structure_assignment.base = base or 50000
 	salary_structure_assignment.variable = 5000
 
 	if not from_date:
